Perf: Avoid cloning listeners in data event emitter
In the core event emitter class for the `@data` module we have a note
explaining that we clone the set of listeners to avoid calling newly
created listeners during the `notifyListeners` loop.

In this patch we use a special token inserted at the end of the listeners
set that indicates when we've reached the last previously-existing listener
while iterating. This is a result of `Set` iteration following insertion
order.

By adding, watching for, and removing this token we can safely iterate
the set without cloning on every call, which can reduce pressure on the
garbage collector and lead to lower memory use and reduced GC overhead.

The results of this change may or may not be easily measurable, but if it's
found to be a positive change regardless it might still be worth introducing
since this code is in the critical path of all state updates, and it runs
following every key press.